+++ Modules/Numerics/Optimizers/src/itkParticleSwarmOptimizer.cxx
@@ -48,7 +48,7 @@ ParticleSwarmOptimizer::UpdateSwarm()
   const itk::Statistics::MersenneTwisterRandomVariateGenerator::Pointer randomGenerator =
     Statistics::MersenneTwisterRandomVariateGenerator::GetInstance();
 
-  const auto n = static_cast<unsigned int>((GetCostFunction())->GetNumberOfParameters());
+  const auto n = GetCostFunction()->GetNumberOfParameters();
 
   for (unsigned int j = 0; j < m_NumberOfParticles; ++j)
   {
